refactor(Comment): 컴포넌트에서 사용하던 Types 분리
