+++ src/components/common/Comment/index.tsx
@@ -2,30 +2,7 @@ import styled from '@emotion/styled';
 import { Text } from '~/components/atom';
 import CommentForm from './CommentForm';
 import CommentItem from './CommentItem';
-
-interface IUser {
-  id: number;
-  nickname: string;
-  profileImage: string;
-}
-
-export interface IRecomment {
-  id: number;
-  comment: string;
-  user: IUser;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface IComment {
-  id: number;
-  comment: string;
-  user: IUser;
-  createdAt: string;
-  updatedAt: string;
-  isDeleted: boolean;
-  recomments: IRecomment[];
-}
+import { IComment } from './types';
 
 const dummyComments = [
   {
