+++ src/components/common/Comment/types.ts
@@ -1,0 +1,23 @@
+interface IUser {
+  id: number;
+  nickname: string;
+  profileImage: string;
+}
+
+export interface IRecomment {
+  id: number;
+  comment: string;
+  user: IUser;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface IComment {
+  id: number;
+  comment: string;
+  user: IUser;
+  createdAt: string;
+  updatedAt: string;
+  isDeleted: boolean;
+  recomments: IRecomment[];
+}
